tts_native/requirements.txt
grpcio==1.75.0
protobuf==6.32.1
prometheus-client==0.21.0
aiohttp==3.10.5
TTS==0.22.0

